Update Email verification endpoint
