+++ package.json
@@ -30,7 +30,8 @@
     "jsonwebtoken": "^9.0.2",
     "mysql": "^2.18.1",
     "mysql2": "^3.6.5",
-    "nodemailer": "^6.9.7"
+    "nodemailer": "^6.9.7",
+    "speakeasy": "^2.0.0",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
