+++ src/routes/verify.route.js
@@ -2,5 +2,6 @@ import { verifyEmail } from "../controllers/verify.controller.js";
 
 
 export default (router) => {
-	router.post('/verify-email', verifyEmail);
+	router.post('/user/verify-email', verifyEmail);
+	// route to request a new OTP
 }
